Бинарный поиск теперь нормальный, созданы классы базовой задачи и функций
