+++ lab/lab2/Zero.java
@@ -1,0 +1,4 @@
+package lab2;
+
+public class Zero {
+}
